test: car 로직 분리로 인한 코드 수정
